+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 
 # Release Notes
 
+## Version 0.16.1: March 18th, 2024
+
+### New Features
+
+- N/A
+
+### Enhancements
+
+- UI improvements to the Sara status messages
+- Encourage Sara NOT to use the specific vector store filenames in her answers to the user - and instead use more general resource names like "Software Architectural Blueprint" - which are names also in the raw resources uploaded by Backend
+- Set the Sara "Deployment" stage in the Assistant id for separation of data
+
+### Bug Fixes
+
+- N/A
+
 ## Version 0.16.0: March 15th, 2024
 
 ### New Features